Cambios finales

+++ base/funciones/alm.ft_preingreso_ime.sql
@@ -31,6 +31,11 @@ DECLARE
 	v_mensaje_error         text;
 	v_id_preingreso			integer;
 	v_result				varchar;
+    v_id_proceso_wf			integer;
+    v_id_estado_wf			integer;
+    v_id_tipo_proceso		integer;
+    v_id_tipo_estado 		integer;
+    v_id_estado_actual  	integer;
 			    
 BEGIN
 
@@ -100,6 +105,19 @@ BEGIN
 	elsif(p_transaccion='SAL_PREING_MOD')then
 
 		begin
+        
+        	--Validación de existencia del registro
+            if not exists(select 1 from alm.tpreingreso
+            			where id_preingreso = v_parametros.id_preingreso) then
+            	raise exception 'Preingreso inexistente';
+            end if;
+            
+            if not exists(select 1 from alm.tpreingreso
+            			where id_preingreso = v_parametros.id_preingreso
+                        and estado = 'borrador') then
+            	raise exception 'El Preingreso debe estar en Borrador';
+            end if;
+            
 			--Sentencia de la modificacion
 			update alm.tpreingreso set
 			id_cotizacion = v_parametros.id_cotizacion,
@@ -134,6 +152,19 @@ BEGIN
 	elsif(p_transaccion='SAL_PREING_ELI')then
 
 		begin
+        
+        	--Validación de existencia del registro
+            if not exists(select 1 from alm.tpreingreso
+            			where id_preingreso = v_parametros.id_preingreso) then
+            	raise exception 'Preingreso inexistente';
+            end if;
+            
+            if not exists(select 1 from alm.tpreingreso
+            			where id_preingreso = v_parametros.id_preingreso
+                        and estado = 'borrador') then
+            	raise exception 'El Preingreso debe estar en Borrador';
+            end if;
+            
 			--Sentencia de la eliminacion
 			delete from alm.tpreingreso
             where id_preingreso=v_parametros.id_preingreso;
@@ -158,11 +189,84 @@ BEGIN
 
 		begin
         
+        	--Validación de existencia del registro
+            if not exists(select 1 from alm.tpreingreso
+            			where id_preingreso = v_parametros.id_preingreso) then
+            	raise exception 'Preingreso inexistente';
+            end if;
+        
 			--Llamada a la función de generación de ingreso
 			v_result = alm.f_generar_ingreso(p_id_usuario, v_parametros.id_preingreso);
                
 			--Definicion de la respuesta
             v_resp = pxp.f_agrega_clave(v_resp,'mensaje','Detalle Preingreso modificado(a)'); 
+               
+            --Devuelve la respuesta
+            return v_resp;
+            
+		end;
+        
+    /*********************************    
+ 	#TRANSACCION:  'SAL_PREING_REV'
+ 	#DESCRIPCION:	Reversión del Preingreso
+ 	#AUTOR:			RCM
+ 	#FECHA:			21/10/2013
+	***********************************/
+
+	elsif(p_transaccion='SAL_PREING_REV')then
+
+		begin
+        	
+        	--Validación de existencia del registro
+            if not exists(select 1 from alm.tpreingreso
+            			where id_preingreso = v_parametros.id_preingreso) then
+            	raise exception 'Preingreso inexistente';
+            end if;
+            
+            if not exists(select 1 from alm.tpreingreso
+            			where id_preingreso = v_parametros.id_preingreso
+                        and estado = 'borrador') then
+            	raise exception 'El Preingreso debe estar en Borrador';
+            end if;
+            
+             --Obtiene el Proceso WF
+            SELECT
+            pre.id_estado_wf, pw.id_tipo_proceso, pw.id_proceso_wf
+           	into
+            v_id_estado_wf, v_id_tipo_proceso, v_id_proceso_wf
+           	FROM alm.tpreingreso pre
+           	inner join wf.tproceso_wf pw on pw.id_proceso_wf = pre.id_proceso_wf
+           	WHERE pre.id_preingreso = v_parametros.id_preingreso;
+            
+            --Obtiene el estado cancelado del WF
+            select 
+            te.id_tipo_estado
+            into
+            v_id_tipo_estado
+            from wf.tproceso_wf pw 
+            inner join wf.ttipo_proceso tp on pw.id_tipo_proceso = tp.id_tipo_proceso
+            inner join wf.ttipo_estado te on te.id_tipo_proceso = tp.id_tipo_proceso and te.codigo = 'cancelado'               
+            where pw.id_proceso_wf = v_id_proceso_wf;
+               
+            --Se cancela el WF
+            v_id_estado_actual =  wf.f_registra_estado_wf(v_id_tipo_estado, 
+                                                           NULL, 
+                                                           v_id_estado_wf, 
+                                                           v_id_proceso_wf,
+                                                           p_id_usuario,
+                                                           null);
+            
+            --Cancela el movimiento
+            update alm.tpreingreso set
+            id_estado_wf =  v_id_estado_actual,
+          	estado = 'cancelado',
+            id_usuario_mod=p_id_usuario,
+            fecha_mod=now()
+        	where id_preingreso = v_parametros.id_preingreso;
+               
+			--Definicion de la respuesta
+            v_resp = pxp.f_agrega_clave(v_resp,'mensaje','Detalle Preingreso modificado(a)'); 
+            v_resp=pxp.f_agrega_clave(v_resp,'id_preingreso',v_parametros.id_preingreso::varchar);
                
             --Devuelve la respuesta
             return v_resp;
